More Changes Based on Using More Repos (#12)
# Summary of PR changes
Changes in order to support more dotnet workflows:
- I found a workflow, [Household Management Dotnet
CI](https://github.com/im-client/household-management/blob/main/.github/workflows/im-build-dotnet-ci.yml#L43),
which has multiple versions of dotnet declared in the value. So changes
had to be made in the setup wizard to be able to read this.
- Adding new library that can read the csharp yml template and convert
it to Json; this will make it easier to read the `DOTNET_VERSION` value.
- Adding `DOTNET_INSTALL_DIR` env variable that is used by some teams in
the `setup-dotnet` action step.
- Adding `body` to PR so it will have an explanation and material that
can be looked at to help teams understand what GHAS does
- Adding configurable ticket url in case we create a custom ticket type
for ITHD requests related to the workflow.
- When running `start` we will check if the repos have code scanning yml
file on them by doing a code search before getting started; if they do
have it the repo will be skipped. This will allow us to run this
codebase periodically to add code scanning to any new repos without
affecting the others.
- Add default values for dotnet job. This will allow creation to succeed
if the Github repo is not using github action workflows yet.
- Remove shared values from `repos.json` and read them at runtime
instead.